third_party/WebKit/LayoutTests/platform/win/external/wpt/IndexedDB/idbcursor_continue_index2-expected.txt

This is a testharness.js-based test.
PASS IDBCursor.continue() - index - attempt to pass a key parameter that is not a valid key 
Harness: the test ran to completion.
